+++ src/main.c
@@ -1,4 +1,4 @@
-#include "menu/binary_system/binary_system.h"
+#include "menu/binary/binary_system.h"
 #include "menu/main/menu_main.h"
 #include "utils/utils.h"
