+++ src/menu/binary/binary_to_decimal.c
