+++ src/menu/binary/binary_validate.c
