+++ src/menu/binary/binary_validate.h
